working on parser

+++ src/main/scala/com/github/davidbailey/TransportSim/models.scala
@@ -12,7 +12,7 @@ import scala.collection.mutable.ListBuffer
   }
 
   case class Point(lat: BigDecimal, lon: BigDecimal) extends Serializable
-  case class Node(id: Int, point: Point) extends Serializable // OSM: Node
+  case class Node(id: BigInt, point: Point) extends Serializable // OSM: Node
 
   class Person (r: List[Point], dt: Int) extends Serializable { // basic agent
     val id = abs(Random.nextInt)
@@ -138,7 +138,7 @@ import scala.collection.mutable.ListBuffer
     val width = 4.0
   }
 
-  case class Way(id: Int, centerLine: List[Node]) extends Serializable // OSM: Way
+  case class Way(id: BigInt, centerLine: List[Node]) extends Serializable // OSM: Way
 
   abstract class HalfRoad extends Serializable { // OSM: Way w/ oneway != yes
     val lanes: List[Lane]
